+++ src/Request/Direct.php
@@ -69,8 +69,12 @@ class Direct extends RequestCollection
         if ($filter !== null) {
             $request->addParam('fetch_reason', $fetchReason);
         }
-        if ($fetchReason === 'initial_snapshot') {
-            $request->addParam('batch_size', 10);
+        $batchSize = $this->ig->getExperimentParam('59489', 1, null);
+        if ($batchSize !== null) {
+            $request->addParam('batch_size', $batchSize);
+        }
+        if ($this->ig->isExperimentEnabled('45863', 0, false)) {
+            $request->addParam('no_pending_badge', 'true');
         }
         if ($seqId !== null) {
             $request->addParam('seq_id', $seqId);